Added export functionality
Now, the programs allows the user to export data in the AVL tree to a file.
