+++ main.cpp
@@ -1,28 +1,8 @@
 #include "AVLTree.h"
+#include "ImportStockData.h"
 #include <iostream>
 #include <iomanip>
 #include <string>
-
-#include <cpprest/http_client.h>
-#include <cpprest/filestream.h>
-#include <cpprest/asyncrt_utils.h>
-
-using namespace web;
-using namespace web::http;
-using namespace web::http::client;
-using namespace web::json;
-
-std::string GET_API_KEY(const std::string& filePath) {
-    std::ifstream config(filePath);
-    std::string API_KEY;
-
-    std::getline(config, API_KEY);
-	API_KEY = API_KEY.substr(8); // Remove "API_KEY=" prefix
-
-	config.close();
-
-	return API_KEY;
-}
 
 void displayMenu() {
     std::cout << "===== Stock Market Data Analyzer =====\n";
@@ -33,8 +13,9 @@ void displayMenu() {
     std::cout << "5. Display all stocks\n";
     std::cout << "6. Display stocks by ticker\n";
     std::cout << "7. Display stocks by date range\n";
-    std::cout << "8. Exit\n";
-    std::cout << "Enter your choice (1-8): ";
+    std::cout << "8. Export to File\n";
+    std::cout << "9. Exit\n";
+    std::cout << "Enter your choice (1-9): ";
 }
 
 StockData inputStockData() {
@@ -65,54 +46,23 @@ void displayStock(const StockData& stock) {
     std::cout << " | Volume: " << stock.volume << "\n";
 }
 
-void importData(AVLTree &stockTree, std::string ticker) {
-    // Fetch from config file
-    std::string configData = GET_API_KEY("config.txt");
-    char* API_KEY = new char[configData.length() + 1];
-    std::strcpy(API_KEY, configData.c_str());
-
-    char* symbol = new char[ticker.length() + 1];
-    std::strcpy(symbol, ticker.c_str());
-    
-    // Create the URI
-    uri_builder builder(U("https://api.stockdata.org/v1/data/eod"));
-    builder.append_query(U("symbols"), symbol);
-    builder.append_query(U("api_token"), API_KEY);
-    builder.append_query(U("date_from"), ("2025-01-01")); // Date to get stocks from onward.
-
-    // Make the HTTP request
-    http_client client(builder.to_uri());
-
-    client.request(methods::GET).then([&stockTree, &ticker](http_response response) {
-        if (response.status_code() == status_codes::OK) {
-            // Parse the JSON response
-            json::value json_response = response.extract_json().get();
-
-            auto data = json_response[U("data")].as_array();
-            if (data.size() != 0) {
-                for (auto& entry : data) {
-                    std::string dateTime = utility::conversions::to_utf8string(entry[U("date")].as_string());
-                    std::string date = dateTime.substr(0, dateTime.find('T')); // Extract only YYYY-MM-DD
-
-                    double open = entry[U("open")].as_double();
-                    double close = entry[U("close")].as_double();
-                    double high = entry[U("high")].as_double();
-                    double low = entry[U("low")].as_double();
-                    int volume = entry[U("volume")].as_integer();
-
-                    StockData temp(ticker, date, open, close, high, low, volume);
-                    stockTree.insert(temp);
-                }
-                std::cout << "Stock data inserted successfully.\n";
-            } 
-            else {
-                std::cout << "No stock data found.\n" << std::endl;
-            }
+void exportDataAsCSV(AVLTree stockTree) {
+    std::ofstream outFile("stocks.csv");
+    if (outFile.is_open()) {
+        outFile << "Ticker,Date,Open,Close,High,Low,Volume\n";
+        std::vector<StockData> allStocks = stockTree.getAllStocks();
+        for (const auto& stock : allStocks) {
+            outFile << stock.ticker << "," << stock.date << ","
+                << stock.openPrice << "," << stock.closePrice << ","
+                << stock.highPrice << "," << stock.lowPrice << ","
+                << stock.volume << "\n";
         }
-        else {
-            std::cerr << "Error: " << response.status_code() << std::endl;
-        }
-        }).wait();
+        outFile.close();
+        std::cout << "Data exported to stocks.csv successfully.\n";
+    }
+    else {
+        std::cout << "Error opening file for export.\n";
+    }
 }
 
 int main() {
@@ -212,22 +162,22 @@ int main() {
             }
             break;
         }
-
         case 8:
+			exportDataAsCSV(stockTree);
+			break;
+        case 9:
+			std::cout << "Exiting the program...\n";
             return 0;
         default:
             std::cout << "Invalid choice. Please try again.\n";
         }
 
-        char temp;
-        std::cin.get(temp);
-
-        // Clear input buffer
-        std::cin.clear();
+        // Wait for user input before clearing the screen
         std::cin.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+        std::cin.get();
 
-		system("cls"); // Clear console
+        // Clear the console screen
+		system("cls");
     }
-
     return 0;
 }
